Add files via upload
We will use real-life datasets to demonstrate the basics of machine learning, which
include preprocessing data for machine learning models and building a regression
model. I used Linear Regression and Random Forest Regression.
We will use a wine dataset
(https://archive.ics.uci.edu/ml/machine-learning-databases/wine/). These data are
the results of a chemical analysis of wines grown in the same region in Italy but
derived from three different cultivars. The analysis determined the quantities of 13
constituents found in each of the three types of wines.